[BUGFIX] Avoid copying workspace delete placeholder
The tests replay the reported scenario: A live content
element with two images is manipulated in workspaces
by deleting one attached image (creating a delete
placeholder, t3ver_state 2) and adding another one
(creating a workspace-new record t3ver_state 1).
The content element is then copied in workspaces.

The tests verify the delete placeholder is ignored
and copied along with the other attached images,
when the fix within DataHandler is not applied.
This makes the deleted attached image magically
pop up again in the copied content element.

The patch fixes this with a straight condition
within DataHandler copyRecord() to suppress
creating copies of records that have a delete
placeholder overlay in this workspace.

Resolves: #61234
Related: #105698
Releases: main, 13.4, 12.4
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/87237
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>

+++ typo3/sysext/core/Classes/DataHandling/DataHandler.php
@@ -3576,6 +3576,13 @@ class DataHandler
         $data = [];
         $nonFields = array_unique(GeneralUtility::trimExplode(',', 'uid,perms_userid,perms_groupid,perms_user,perms_group,perms_everybody,t3ver_oid,t3ver_wsid,t3ver_state,t3ver_stage,' . $excludeFields, true));
         BackendUtility::workspaceOL($table, $row, $this->BE_USER->workspace);
+        if (BackendUtility::isTableWorkspaceEnabled($table)
+            && $this->BE_USER->workspace > 0
+            && VersionState::tryFrom($row['t3ver_state'] ?? 0) === VersionState::DELETE_PLACEHOLDER
+        ) {
+            // The to-copy record turns out to be a delete placeholder. Those do not make sense to be copied and are skipped.
+            return null;
+        }
         $row = BackendUtility::purgeComputedPropertiesFromRecord($row);
 
         // Initializing:
